move string_utils string directory in utils directory
