+++ adapter/internal/utils/string/string_utils.go
