+++ adapter/internal/utils/string/string_utils_test.go
